code clean up

+++ src/map.h
@@ -199,7 +199,7 @@ class _Map {
 
 		void ChangeMapState(const _Event *Event);
 		bool CanChangeMapState(const _Event *Event);
-		void ToggleEventActive(int Index);
+		void ToggleEventActive(size_t Index);
 		void SwapBlockTextures(int Layer, int Index);
 		bool HasEvents(const glm::ivec2 &Position) const;
